Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main() {
     int r;
-    float area;
+    double area;
     const float pi = 3.14;
 
     scanf("%d", &r);
     area = pi * r * r;
-    printf("Area: %.2f", area);
+    printf("Area: %.2lf", area);
 
 
     return 0;
